seperated peoplelistsidebar as an independent componenent
