+++ src/routes/Groups/Components/GroupsView.js
@@ -2,82 +2,23 @@
  * Created by rui on 4/7/17.
  */
 import React from 'react'
-import { Label, Segment, Grid, Image, List } from 'semantic-ui-react'
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import { Input, Label, Segment, Grid, Image, List } from 'semantic-ui-react'
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
 import { Drawer } from "material-ui"
+import Sticky from "react-stickynode"
+
+import PeopleListSidebar from "./PeopleListSidebar"
 
 import "./GroupsView.scss"
 import GroupCard from "./GroupCard/GroupCard"
 
-class Person extends React.Component {
-    constructor(props) {
-        super(props)
-    }
-    render() {
-        return (
-            <List.Item>
-                <Image avatar src={ this.props.image }/>
-                <List.Content>
-                    <List.Header> { this.props.name } </List.Header>
-                </List.Content>
-                <List.Content floated="right" verticalAlign="middle">
-                    <Label color={ (this.props.isAssigned) ? "green" : "red" }>
-                        { (this.props.isAssigned) ? "assigned" : "unassigned"}
-                    </Label>
-                </List.Content>
-            </List.Item>
-        )
-    }
-}
-
-class PeopleListSidebar extends React.Component {
-    constructor(props) {
-        super(props)
-    }
-
-    render() {
-        return (
-            <MuiThemeProvider>
-                <Drawer
-                    docked={ true }
-                    open={ true }
-                    zDepth={ 1 }
-                    style = { drawerStyle }
-                    overlayStyle={ drawerStyle }
-                >
-                    <Segment basic>
-                        <List animated verticalAlign='middle' size="large" selection>
-                            {
-                                people.map((personObj) => (
-                                        <Person
-                                            name = { personObj.name }
-                                            isAssigned = { personObj.isAssigned }
-                                            image = { personObj.image }
-                                        />
-                                    )
-                                )
-                            }
-                        </List>
-                    </Segment>
-                </Drawer>
-            </MuiThemeProvider>
-        )
-    }
-
-}
-
-Person.propTypes = {
-     isAssigned: React.PropTypes.bool.isRequired,
-     name: React.PropTypes.string.isRequired,
-     image: React.PropTypes.string.isRequired
-}
 
 // test
 let person =
     {
         name: "rui tu",
         image: "http://react.semantic-ui.com/assets/images/avatar/small/joe.jpg",
-        isAssigned: true
+        groupNumber: -1
     }
 let people = [ ]
 
@@ -93,9 +34,6 @@ for (let i = 0; i < 19; i++ ) {
     groups.push( { groupNumber: 1, capacity: 8, members: people.slice(0, getRandomArbitrary(0, 10)) } )
 }
 
-const drawerStyle = {
-    zIndex: 0
-}
 
 export class GroupsView extends React.Component {
     constructor(props) {
@@ -105,7 +43,6 @@ export class GroupsView extends React.Component {
     render() {
         const peopleListWidth = 4;
         const groupCardsWidth = 12;
-        const peopleListTopPadding = 110;
 
         let getGroups = (groups) => {
             return (
@@ -123,7 +60,7 @@ export class GroupsView extends React.Component {
 
         return (
             <div>
-            <PeopleListSidebar />
+            <PeopleListSidebar people={ people } />
                 <div className="container">
                     <Grid >
                         <Grid.Row>
