+++ src/routes/Groups/Components/PeopleListSidebar/index.js
@@ -1,0 +1,6 @@
+/**
+ * Created by rui on 4/10/17.
+ */
+import PeopleListSidebar from "./PeopleListSidebar"
+
+export default PeopleListSidebar
